add ProductDetail class (DTO) - changing Double to BigDecimal

+++ src/main/java/pl/sdacademy/pewex/product/model/ProductDetail.java
@@ -17,7 +17,7 @@ public class ProductDetail {
     private String title;
     private BigDecimal price;
     private String description;
-    private Double rating;
+    private BigDecimal rating;
